// src/plugins/elasticsearch_ui_shared/static/forms/hook_form_lib/types.ts
export interface FieldConfig<T = FormData> {
  readonly path?: string;
  readonly label?: string;
  readonly helpText?: string;
  readonly type?: HTMLInputElement['type'];
  readonly defaultValue?: unknown;
  readonly validations?: Array<ValidationConfig<T>>;
  readonly formatters?: FormatterFunc[];
  readonly deSerializer?: SerializerFunc;
  readonly serializer?: SerializerFunc;
  readonly fieldsToValidateOnChange?: string[];
  readonly isValidationAsync?: boolean;
  readonly errorDisplayDelay?: number;
}

export interface FieldsMap {
  [key: string]: Field;
}

export type FormSubmitHandler<T> = (formData: T, isValid: boolean) => Promise<void>;

export interface ValidationError {
  message: string | ((error: ValidationError) => string);
  code?: string;
  validationType?: string;
  [key: string]: any;
}

export type ValidationFunc<T = any> = (data: {
  path: string;
  value: unknown;
  form: Form<T>;
  formData: T;
  errors: readonly ValidationError[];
}) => ValidationError | void | undefined | Promise<ValidationError | void | undefined>;

export interface FieldValidateResponse {
  isValid: boolean;
  errors: ValidationError[];
}

export type SerializerFunc<T = unknown> = (value: any) => T;

export type FormData = Record<string, unknown>;

type FormatterFunc = (value: any) => unknown;

// We set it as unknown as a form field can be any of any type
// string | number | boolean | string[] ...
type FieldValue = unknown;

export interface ValidationConfig<T = any> {
  validator: ValidationFunc<T>;
  type?: string;
  exitOnFail?: boolean;
}